+++ cedinia/src/model.rs
@@ -1,7 +1,7 @@
-use slint::{Model, ModelRc, SharedString, VecModel};
+use slint::{ComponentHandle, Model, ModelRc, SharedString, VecModel};
 
-use crate::FileEntry;
 use crate::scan_runner::FileItem;
+use crate::{AppState, FileEntry, MainWindow, SimilarGroupCard, SimilarImageItem};
 
 pub fn make_file_model(items: Vec<FileItem>) -> ModelRc<FileEntry> {
     let entries: Vec<FileEntry> = items
@@ -32,6 +32,107 @@ pub fn toggle_row(model: &ModelRc<FileEntry>, index: usize) {
 
 pub fn count_checked(model: &ModelRc<FileEntry>) -> i32 {
     model.iter().filter(|e: &FileEntry| e.checked).count() as i32
+}
+
+pub(crate) fn build_gallery_groups(items: &[FileItem], placeholder: &slint::Image) -> Vec<SimilarGroupCard> {
+    use slint::{ModelRc, SharedString, VecModel};
+
+    use crate::common::{STR_IDX_NAME, STR_IDX_PATH, STR_IDX_SIZE};
+    let mut groups: Vec<SimilarGroupCard> = Vec::new();
+    let mut cur_label = String::new();
+    let mut cur_items: Vec<SimilarImageItem> = Vec::new();
+
+    for (flat_idx, item) in items.iter().enumerate() {
+        if item.is_header {
+            if !cur_items.is_empty() {
+                groups.push(SimilarGroupCard {
+                    label: SharedString::from(&cur_label),
+                    items: ModelRc::new(VecModel::from(std::mem::take(&mut cur_items))),
+                });
+            }
+            cur_label = item.val_str[STR_IDX_NAME].clone();
+        } else {
+            let name = &item.val_str[STR_IDX_NAME];
+            let path = &item.val_str[STR_IDX_PATH];
+            let size = &item.val_str[STR_IDX_SIZE];
+            let full_path = if path.is_empty() { name.clone() } else { format!("{path}/{name}") };
+            cur_items.push(SimilarImageItem {
+                full_path: SharedString::from(full_path),
+                name: SharedString::from(name),
+                size: SharedString::from(size),
+                val_str: ModelRc::new(VecModel::from(item.val_str.iter().map(|s| SharedString::from(s.as_str())).collect::<Vec<_>>())),
+                flat_idx: flat_idx as i32,
+                thumbnail: placeholder.clone(),
+                checked: false,
+                is_reference: item.is_reference,
+            });
+        }
+    }
+    if !cur_items.is_empty() {
+        groups.push(SimilarGroupCard {
+            label: SharedString::from(&cur_label),
+            items: ModelRc::new(VecModel::from(cur_items)),
+        });
+    }
+    groups
+}
+
+pub(crate) fn rebuild_similar_images_after_delete(win: &MainWindow, deleted: &std::collections::HashSet<String>) {
+    let groups = win.get_similar_images_groups();
+    let mut new_groups: Vec<SimilarGroupCard> = Vec::new();
+    let mut new_flat: Vec<FileEntry> = Vec::new();
+
+    for gi in 0..groups.row_count() {
+        if let Some(group) = groups.row_data(gi) {
+            let surviving: Vec<_> = (0..group.items.row_count())
+                .filter_map(|ii| group.items.row_data(ii))
+                .filter(|item| !deleted.contains(item.full_path.as_str()))
+                .map(|mut item| {
+                    item.checked = false;
+                    item
+                })
+                .collect();
+
+            if surviving.is_empty() {
+                continue;
+            }
+
+            new_flat.push(FileEntry {
+                checked: false,
+                is_header: true,
+                is_reference: false,
+                val_str: ModelRc::new(VecModel::from(vec![
+                    group.label.clone(),
+                    SharedString::default(),
+                    SharedString::default(),
+                    SharedString::default(),
+                ])),
+                val_int: ModelRc::new(VecModel::from(Vec::new())),
+            });
+
+            let mut final_items: Vec<SimilarImageItem> = Vec::new();
+            for mut item in surviving {
+                item.flat_idx = new_flat.len() as i32;
+                new_flat.push(FileEntry {
+                    checked: false,
+                    is_header: false,
+                    is_reference: false,
+                    val_str: item.val_str.clone(),
+                    val_int: ModelRc::new(VecModel::from(Vec::new())),
+                });
+                final_items.push(item);
+            }
+
+            new_groups.push(SimilarGroupCard {
+                label: group.label.clone(),
+                items: ModelRc::new(VecModel::from(final_items)),
+            });
+        }
+    }
+
+    win.set_similar_images_model(ModelRc::new(VecModel::from(new_flat)));
+    win.set_similar_images_groups(ModelRc::new(VecModel::from(new_groups)));
+    win.global::<AppState>().set_selected_count(0);
 }
 
 #[cfg(test)]
