+++ czkawka_core/src/common/path_utils.rs
@@ -1,0 +1,88 @@
+use std::cmp::Ordering;
+use std::ffi::OsString;
+use std::path::{Path, PathBuf};
+
+use crate::common::items::SingleExcludedItem;
+
+pub fn split_path(path: &Path) -> (String, String) {
+    match (path.parent(), path.file_name()) {
+        (Some(dir), Some(file)) => (dir.to_string_lossy().to_string(), file.to_string_lossy().into_owned()),
+        (Some(dir), None) => (dir.to_string_lossy().to_string(), String::new()),
+        (None, _) => (String::new(), String::new()),
+    }
+}
+
+pub fn split_path_compare(path_a: &Path, path_b: &Path) -> Ordering {
+    match path_a.parent().cmp(&path_b.parent()) {
+        Ordering::Equal => path_a.file_name().cmp(&path_b.file_name()),
+        other => other,
+    }
+}
+
+#[expect(clippy::string_slice)]
+#[expect(clippy::indexing_slicing)]
+pub fn regex_check(expression_item: &SingleExcludedItem, directory_name: &str) -> bool {
+    if expression_item.expression_splits.is_empty() {
+        return true;
+    }
+
+    // Early checking if directory contains all parts needed by expression
+    for split in &expression_item.unique_extensions_splits {
+        if !directory_name.contains(split) {
+            return false;
+        }
+    }
+
+    // `git*` shouldn't be true for `/gitsfafasfs`
+    if !expression_item.expression.starts_with('*')
+        && directory_name
+            .find(&expression_item.expression_splits[0])
+            .expect("Cannot fail, because split must exists in directory_name")
+            > 0
+    {
+        return false;
+    }
+    // `*home` shouldn't be true for `/homeowner`
+    if !expression_item.expression.ends_with('*')
+        && !directory_name.ends_with(expression_item.expression_splits.last().expect("Cannot fail, because at least one item is available"))
+    {
+        return false;
+    }
+
+    // At the end we check if parts between * are correctly positioned.
+    // Cursor tracks the byte position immediately after the previously matched split;
+    // each subsequent split must be found at or after that cursor.
+    let first = &expression_item.expression_splits[0];
+    let first_pos = directory_name.find(first).expect("Cannot fail, because is checked earlier");
+    let mut cursor = first_pos + first.len();
+    for spl in &expression_item.expression_splits[1..] {
+        let Some(rel) = directory_name[cursor..].find(spl) else { return false };
+        cursor += rel + spl.len();
+    }
+    true
+}
+
+#[expect(clippy::string_slice)] // Is in char boundary
+pub fn normalize_windows_path<P: AsRef<Path>>(path_to_change: P) -> PathBuf {
+    let path = path_to_change.as_ref();
+
+    // Don't do anything, because network path may be case intensive
+    if path.to_string_lossy().starts_with('\\') {
+        return path.to_path_buf();
+    }
+
+    match path.to_str() {
+        Some(path) if path.is_char_boundary(1) => {
+            let replaced = path.replace('/', "\\");
+            let mut new_path = OsString::new();
+            if replaced[1..].starts_with(':') {
+                new_path.push(replaced[..1].to_ascii_uppercase());
+                new_path.push(replaced[1..].to_ascii_lowercase());
+            } else {
+                new_path.push(replaced.to_ascii_lowercase());
+            }
+            PathBuf::from(new_path)
+        }
+        _ => path.to_path_buf(),
+    }
+}
